requirements.app.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
httpx==0.28.1
prometheus-client==0.21.0
kafka-python==2.0.2
numpy==1.26.4
Pillow==10.4.0
opencv-python-headless==4.10.0.84
scipy==1.14.1
scikit-image==0.24.0
onnx==1.21.0
onnxruntime==1.20.1
ultralytics==8.3.39
nibabel==5.3.2
torch==2.3.1
torchvision==0.18.1
monai==1.4.0
